requirements.txt
Flask==1.1.2
mysqlclient==2.0.3
flask-mysqldb==0.2.0
